requirements.txt
# Core dependencies
openai-whisper>=20231117
pyttsx3>=2.90
pyaudio>=0.2.11
langchain-community>=0.0.10
ollama>=0.1.7
wave

# Optional: For better audio processing
soundfile>=0.12.1
librosa>=0.10.1

# Optional: For enhanced LangChain features
langchain>=0.1.0
langchain-core>=0.1.0